Publish snapshot releases

+++ .circleci/publish-cdn.sh
@@ -3,8 +3,11 @@
 if [[ -v CIRCLE_TAG ]]; then
     echo "Deploying CDN updates for $CIRCLE_TAG"
 else
-    echo "CDN updates are not published for untagged builds"
-    exit
+    if [ "$CIRCLE_PROJECT_USERNAME" != "docbook" \
+         -o "$CIRCLE_BRANCH" != "main" ]; then
+        echo "Snapshots are only for docbook/main commits"
+        exit
+    fi
 fi
 
 if [ `git branch -r | grep "origin/gh-pages" | wc -l` = 0 ]; then
@@ -32,16 +35,21 @@ mkdir cdn
 cd cdn
 git clone --depth 1 -c core.sshCommand="/usr/bin/ssh -i $IDRSA" git@github.com:docbook/cdn.git .
 
-# We should never be republishing to the same tag, but just in case...
-rm -rf release/xsltng/$CIRCLE_TAG
-mkdir -p release/xsltng/$CIRCLE_TAG
-rsync -ar --exclude bin --exclude libs --exclude samples --exclude docker \
-      ../repo/build/stage/zip/ release/xsltng/$CIRCLE_TAG/
+if [[ -v CIRCLE_TAG ]]; then
+    mkdir -p release/xsltng/snapshot
+    cp ../repo/build/distributions/*.zip release/xsltng/snapshot/
+else
+    # We should never be republishing to the same tag, but just in case...
+    rm -rf release/xsltng/$CIRCLE_TAG
+    mkdir -p release/xsltng/$CIRCLE_TAG
+    rsync -ar --exclude bin --exclude libs --exclude samples --exclude docker \
+          ../repo/build/stage/zip/ release/xsltng/$CIRCLE_TAG/
 
-# Make this the current release too
-rm -rf release/xsltng/current
-mkdir -p release/xsltng/current
-rsync -ar release/xsltng/$CIRCLE_TAG/ release/xsltng/current/
+    # Make this the current release too
+    rm -rf release/xsltng/current
+    mkdir -p release/xsltng/current
+    rsync -ar release/xsltng/$CIRCLE_TAG/ release/xsltng/current/
+fi
 
 # generate indexes
 perl bin/make-indexes.pl release/xsltng
